Update event-loop.js
I try to use nashorn-async to enable promise-7.0.4.min.js for running pdf-lib.js in nashorn. It is almost successful. Than you for nashorn-async.
And I found an error because of calling clearTimeout before setTimeout in promise-7.0.4.min.js. And it is no error in browses.
You can try the next code in browsers.

var a; clearTimeout(a);

So I think it may be good to add a checking in clearTimeout and clearInterval.